+++ src/about/About.tsx
@@ -1,5 +1,5 @@
 import { Link } from 'react-router-dom';
-import Header from './Header';
+import Header from '../utilities/Header';
 
 function About() {
     return (
